+++ src/frontend/components/TranscriptSentence/index.tsx
